Aplikasi E-Commerce Laravel 6 #4: Upload Products

+++ app/Http/Controllers/CategoryController.php
@@ -48,8 +48,8 @@ class CategoryController extends Controller
 
     public function destroy($id)
     {
-        $category = Category::withCount(['child'])->find($id);
-        if ($category->child_count == 0) {
+        $category = Category::withCount(['child', 'product'])->find($id);
+        if ($category->child_count == 0 && $category->product_count == 0) {
             $category->delete();
             return redirect(route('category.index'))->with(['success' => 'Kategori Dihapus!']);
         }
